+++ index.html
@@ -164,6 +164,8 @@
       </div>
     </section>
   </template>
-<script src="/js/main.js"></script>
+<script type="module"src="/js/main.js"></script>
+<script type="module" src="/js/data.js"></script>
+<script type="module" src="/js/util.js"></script>
 </body>
 </html>
