+++ js/util.js
@@ -1,0 +1,19 @@
+//Получаем рандомное число
+const getRandomNumber = (min, max) => {
+  if (min < 0 || max < 0 || min >= max) {
+    return NaN;
+  }
+
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+
+};
+
+//Проверяем максимальную длину строки
+const checkMaxCommentLength = (comment, maxLength) => comment.length <= maxLength;
+
+//Получение случайного элемента массива
+const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
+
+export {getRandomNumber, checkMaxCommentLength, getRandomArrayElement};
